+++ test/sumo/test_modes/test_prepare/test_prepare.py
@@ -1,7 +1,7 @@
 from sklearn.preprocessing import StandardScaler
-from sumo.modes.prepare.prepare import SumoPrepare, filter_features_and_samples, load_data_npz, load_data_text
+from sumo.modes.prepare.prepare import SumoPrepare
 from sumo.constants import PREPARE_DEFAULTS
-from sumo.utils import save_arrays_to_npz, load_npz
+from sumo.utils import load_npz
 from pandas import DataFrame
 import numpy as np
 import os
@@ -20,21 +20,16 @@ def test_init(tmpdir):
     with pytest.raises(AttributeError):
         SumoPrepare()
 
-    npz_infile = os.path.join(tmpdir, "indata.npz")
     fname_outfile = os.path.join(tmpdir, "outdata.npz")
-    args = _get_args([npz_infile], fname_outfile)
+    txt_infile = os.path.join(tmpdir, "indata.txt")
+    args = _get_args([txt_infile], fname_outfile)
 
     # no input file
     with pytest.raises(FileNotFoundError):
         SumoPrepare(**args)
 
-    # one npz input file
-    data = np.random.random((10, 20))
-    save_arrays_to_npz({'f': data}, npz_infile)
-    SumoPrepare(**args)
-
     # one txt input file
-    txt_infile = os.path.join(tmpdir, "indata.txt")
+    data = np.random.random((10, 20))
     df = DataFrame(data, columns=['sample_{}'.format(i) for i in range(data.shape[1])],
                    index=['feature_{}'.format(i) for i in range(data.shape[0])])
     df.to_csv(txt_infile, sep=" ")
@@ -49,19 +44,19 @@ def test_init(tmpdir):
         SumoPrepare(**args)
 
     # two input files
-    args = _get_args([npz_infile, npz_infile], fname_outfile)
+    args = _get_args([txt_infile, txt_infile], fname_outfile)
     SumoPrepare(**args)
 
     # unsupported similarity method
-    args = _get_args([npz_infile], fname_outfile)
+    args = _get_args([txt_infile], fname_outfile)
     args['method'] = ['random']
     with pytest.raises(ValueError):
         SumoPrepare(**args)
 
 
 def test_run(tmpdir):
-    npz_infile_1 = os.path.join(tmpdir, "indata1.npz")
-    npz_infile_2 = os.path.join(tmpdir, "indata2.npz")
+    infile_1 = os.path.join(tmpdir, "indata1.tsv")
+    infile_2 = os.path.join(tmpdir, "indata2.tsv")
     fname_outfile = os.path.join(tmpdir, "outdata.npz")
     logfile = os.path.join(tmpdir, "prepare.log")
     plots = os.path.join(tmpdir, "plot.png")
@@ -71,15 +66,17 @@ def test_run(tmpdir):
     f0 = np.random.random((20, 10))
     f0 = sc.fit_transform(f0.T).T
     samples1 = ['sample_{}'.format(i) for i in range(10)]
-    save_arrays_to_npz({'f': f0, 'samples': np.array(samples1)}, npz_infile_1)
+    df = DataFrame(f0, columns=samples1)
+    df.to_csv(infile_1, sep="\t")
 
     f1 = np.random.random((40, 12))
     f1 = sc.fit_transform(f1.T).T
     samples2 = ['sample_{}'.format(i) for i in range(12)]
-    save_arrays_to_npz({'f': f1, 'samples': np.array(samples2)}, npz_infile_2)
+    df2 = DataFrame(f1, columns=samples2)
+    df2.to_csv(infile_2, sep="\t")
 
     # incorrect number of similarity methods
-    args = _get_args([npz_infile_1, npz_infile_2], fname_outfile)
+    args = _get_args([infile_1, infile_2], fname_outfile)
     args['names'] = 'samples'
     args['logfile'] = logfile
     args['plot'] = plots
@@ -117,8 +114,7 @@ def test_load_all_data(tmpdir):
     data_vals = np.random.random((10, 20))
     data = DataFrame(data_vals.T, columns=['sample_{}'.format(i) for i in range(data_vals.shape[0])],
                      index=['feature_{}'.format(i) for i in range(data_vals.shape[1])])
-    data.to_csv(txt_infile,
-                sep="\t")  # TODO: change to accept tsv (interpret txt as space delimited), change command line doc string
+    data.to_csv(txt_infile, sep=" ")
 
     sc = SumoPrepare(**args)
     matrices = sc.load_all_data()
@@ -126,120 +122,3 @@ def test_load_all_data(tmpdir):
     assert len(matrices) == 1
     assert fname == txt_infile
     assert mat.values.shape == (20, 10)
-
-
-def test_filter_features_and_samples():
-    data_vals = np.random.random((10, 20))
-    data = DataFrame(data_vals.T, columns=['sample_{}'.format(i) for i in range(data_vals.shape[0])],
-                     index=['feature_{}'.format(i) for i in range(data_vals.shape[1])])
-
-    filtered = filter_features_and_samples(data)
-    assert filtered.values.shape == (20, 10)
-
-    # missing samples and features
-    new_data = data.copy()
-    new_data['sample_0'] = np.nan
-    new_data.loc['feature_0'] = np.nan
-
-    filtered = filter_features_and_samples(new_data)
-    assert 'feature_0' not in filtered.index
-    assert 'sample_0' not in filtered.columns
-
-    # missing values sample filtering
-    new_data = data.copy()
-    new_data['sample_0'][1:3] = np.nan
-
-    filtered = filter_features_and_samples(new_data)
-    assert 'sample_0' in filtered.columns
-
-    filtered = filter_features_and_samples(new_data, drop_samples=0.05)
-    assert 'sample_0' not in filtered.columns
-
-    # missing values feature filtering
-    new_data = data.copy()
-    new_data.loc['feature_0'][1] = np.nan
-
-    filtered = filter_features_and_samples(new_data)
-    assert 'feature_0' in filtered.index
-
-    filtered = filter_features_and_samples(new_data, drop_features=0.05)
-    assert 'feature_0' not in filtered.index
-
-
-def test_load_data_text(tmpdir):
-    fname = os.path.join(tmpdir, "data.tsv")
-    with pytest.raises(FileNotFoundError):
-        load_data_text(file_path=fname)
-
-    # empty data
-    empty_data = DataFrame()
-    empty_data.to_csv(fname, sep="\t")
-    with pytest.raises(ValueError):
-        load_data_text(file_path=fname)
-
-    data_vals = np.random.random((10, 20))
-    data = DataFrame(data_vals.T, columns=['sample_{}'.format(i) for i in range(data_vals.shape[0])],
-                     index=['feature_{}'.format(i) for i in range(data_vals.shape[1])])
-
-    # incorrectly formatted data
-    data.to_csv(fname, sep=",")
-    with pytest.raises(ValueError):
-        load_data_text(file_path=fname)
-
-    # non-nummerical values
-    data.to_csv(fname, sep="\t")
-    with pytest.raises(ValueError):
-        load_data_text(file_path=fname)
-
-    # tab delimited file
-    load_data_text(file_path=fname, sample_names=0, feature_names=0)
-
-    # space delimited file
-    data.to_csv(fname, sep=" ")
-    load_data_text(file_path=fname, sample_names=0, feature_names=0)
-
-    # .gz file
-    data.to_csv(fname + ".gz", sep=" ", compression='gzip')
-    load_data_text(file_path=fname, sample_names=0, feature_names=0)
-
-    # .bz2 file
-    data.to_csv(fname + ".bz2", sep=" ", compression='bz2')
-    load_data_text(file_path=fname, sample_names=0, feature_names=0)
-
-
-def test_load_data_npz(tmpdir):
-    fname = os.path.join(tmpdir, "data.npz")
-    with pytest.raises(FileNotFoundError):
-        load_data_npz(file_path=fname)
-
-    # empty data
-    save_arrays_to_npz({}, file_path=fname)
-    with pytest.raises(ValueError):
-        load_data_npz(file_path=fname)
-
-    # no sample names
-    data_vals = np.random.random((10, 20))
-    save_arrays_to_npz({'f': data_vals.T}, file_path=fname)
-    loaded = load_data_npz(file_path=fname)
-    assert len(loaded) == 1
-    assert loaded[0].shape == (20, 10)
-
-    # sample names
-    sample_names = ['sample_{}'.format(i) for i in range(data_vals.shape[0])]
-    save_arrays_to_npz({'f': data_vals.T, 'samples': np.array(sample_names)}, file_path=fname)
-    loaded = load_data_npz(file_path=fname, sample_idx='samples')
-    assert all([loaded[0].columns[i] == sample_names[i] for i in range(len(sample_names))])
-
-    # missing sample names idx
-    with pytest.raises(AttributeError):
-        load_data_npz(file_path=fname)
-
-    # incorrect sample names idx
-    with pytest.raises(ValueError):
-        load_data_npz(file_path=fname, sample_idx='sample_names')
-
-    # incorrect sample names
-    sample_names = ['sample_{}'.format(i) for i in range(100)]
-    save_arrays_to_npz({'f': data_vals.T, 'samples': np.array(sample_names)}, file_path=fname)
-    with pytest.raises(ValueError):
-        load_data_npz(file_path=fname, sample_idx='sample_names')